+++ packages/f-header/vue.config.js
@@ -1,4 +1,7 @@
-const magicImporter = require('node-sass-magic-importer');
+const path = require('path');
+
+const rootDir = path.join(__dirname, '..', '..');
+const sassOptions = require('../../config/sassOptions')(rootDir);
 const responseLoggedIn = require('./src/components/tests/__mocks__/api.account.details.json');
 const responseLoggedOut = require('./src/components/tests/__mocks__/api.account.details.loggedout.json');
 
@@ -11,9 +14,9 @@ module.exports = {
             .use('importer')
             .loader('sass-loader')
             .options({
-                importer: magicImporter(),
+                ...sassOptions,
                 // eslint-disable-next-line quotes
-                data: `@import "@/assets/scss/common.scss";`
+                additionalData: `@import "../assets/scss/common.scss";`
             });
     },
     devServer: {